package structure refactoring
