+++ .gitignore
@@ -3,5 +3,3 @@
 /package-lock.json
 
 /lib
-/index.d.ts
-/index.js
