+++ src/index.ts
@@ -1,3 +1,3 @@
 import * as Client from 'poeditor-client'
 export {Client}
-export {getProject, pullTranslations} from './lib/utils'
+export {getProject, pullTranslations} from './utils'
