+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
 	"compilerOptions": {
-		"outDir": "",
+		"outDir": "./lib",
 		"declaration": true
 	},
 	"include": [
